Remove unused stuff
Remove leftover nier stuff.
