--- bxm/common/approxMapOffsets.py
@@ -1,737 +1,0 @@
-approxMapOffsets = {
-	"g10218": [
-		-1500.216552734375,
-		692.2970581054688,
-		0.0
-	],
-	"g10318": [
-		-1350.2166748046875,
-		605.6943969726562,
-		0.0
-	],
-	"g10319": [
-		-1350.2166748046875,
-		432.48931884765625,
-		0.0
-	],
-	"g10415": [
-		-1200.216552734375,
-		1038.70703125,
-		0.0
-	],
-	"g10416": [
-		-1200.216552734375,
-		865.5020141601562,
-		0.0
-	],
-	"g10417": [
-		-1200.216552734375,
-		692.2969360351562,
-		0.0
-	],
-	"g10418": [
-		-1200.216552734375,
-		519.0918579101562,
-		0.0
-	],
-	"g10419": [
-		-1200.2166748046875,
-		345.88677978515625,
-		0.0
-	],
-	"g10514": [
-		-1050.216552734375,
-		1125.3095703125,
-		0.0
-	],
-	"g10515": [
-		-1050.216552734375,
-		952.1045532226562,
-		0.0
-	],
-	"g10516": [
-		-1050.216552734375,
-		778.8994750976562,
-		0.0
-	],
-	"g10517": [
-		-1050.216552734375,
-		605.6943969726562,
-		0.0
-	],
-	"g10518": [
-		-1050.216552734375,
-		432.48944091796875,
-		0.0
-	],
-	"g10519": [
-		-1050.2166748046875,
-		259.28436279296875,
-		0.0
-	],
-	"g10520": [
-		-1050.2166748046875,
-		86.0792236328125,
-		0.0
-	],
-	"g10614": [
-		-900.2164306640625,
-		1038.70703125,
-		0.0
-	],
-	"g10615": [
-		-900.2164916992188,
-		865.5020141601562,
-		0.0
-	],
-	"g10616": [
-		-900.2164916992188,
-		692.2969360351562,
-		0.0
-	],
-	"g10617": [
-		-900.216552734375,
-		519.0919799804688,
-		0.0
-	],
-	"g10618": [
-		-900.2166137695312,
-		345.88690185546875,
-		0.0
-	],
-	"g10619": [
-		-900.2166748046875,
-		172.68182373046875,
-		0.0
-	],
-	"g10713": [
-		-750.2163696289062,
-		1125.3095703125,
-		0.0
-	],
-	"g10714": [
-		-750.2164306640625,
-		952.1045532226562,
-		0.0
-	],
-	"g10715": [
-		-750.2164916992188,
-		778.8994750976562,
-		0.0
-	],
-	"g10716": [
-		-750.216552734375,
-		605.6943969726562,
-		0.0
-	],
-	"g10717": [
-		-750.216552734375,
-		432.48931884765625,
-		0.0
-	],
-	"g10718": [
-		-750.2166748046875,
-		259.28424072265625,
-		0.0
-	],
-	"g10719": [
-		-750.2167358398438,
-		86.0791015625,
-		0.0
-	],
-	"g10722": [
-		-750.4308471679688,
-		-428.4694519042969,
-		0.0
-	],
-	"g10812": [
-		-600.2164306640625,
-		1211.912109375,
-		0.0
-	],
-	"g10813": [
-		-600.21630859375,
-		1038.707275390625,
-		0.0
-	],
-	"g10814": [
-		-600.2164306640625,
-		865.5020141601562,
-		0.0
-	],
-	"g10815": [
-		-600.2168579101562,
-		692.2975463867188,
-		0.0
-	],
-	"g10816": [
-		-600.0177612304688,
-		519.5239868164062,
-		0.0
-	],
-	"g10817": [
-		-600.2166137695312,
-		345.88677978515625,
-		0.0
-	],
-	"g10818": [
-		-600.2166748046875,
-		172.68170166015625,
-		0.0
-	],
-	"g10820": [
-		-600.4114379882812,
-		-168.7242431640625,
-		0.0
-	],
-	"g10821": [
-		-600.4321899414062,
-		-341.87237548828125,
-		0.0
-	],
-	"g10822": [
-		-600.0316162109375,
-		-514.9127197265625,
-		0.0
-	],
-	"g10912": [
-		-463.0992126464844,
-		1253.018310546875,
-		0.0
-	],
-	"g10914": [
-		-450.103271484375,
-		781.9497680664062,
-		0.0
-	],
-	"g10915": [
-		-450.2168884277344,
-		605.6950073242188,
-		0.0
-	],
-	"g10916": [
-		-450.2155456542969,
-		432.49017333984375,
-		0.0
-	],
-	"g10917": [
-		-450.2691650390625,
-		259.18878173828125,
-		0.0
-	],
-	"g10918": [
-		-450.1297302246094,
-		86.0806884765625,
-		0.0
-	],
-	"g10919": [
-		-450.1297607421875,
-		-87.1243896484375,
-		0.0
-	],
-	"g10920": [
-		-450.4114685058594,
-		-255.326904296875,
-		0.0
-	],
-	"g10921": [
-		-450.3250427246094,
-		-428.4050598144531,
-		0.0
-	],
-	"g10922": [
-		-450.3506164550781,
-		-601.2713623046875,
-		0.0
-	],
-	"g10923": [
-		-450.3504943847656,
-		-774.326171875,
-		0.0
-	],
-	"g10924": [
-		-450.3506164550781,
-		-748.4771118164062,
-		0.0
-	],
-	"g11013": [
-		-301.1095275878906,
-		864.2005004882812,
-		0.0
-	],
-	"g11014": [
-		-301.10955810546875,
-		690.9953002929688,
-		0.0
-	],
-	"g11015": [
-		-288.1788635253906,
-		525.6282348632812,
-		0.0
-	],
-	"g11016": [
-		-300.34954833984375,
-		345.68548583984375,
-		0.0
-	],
-	"g11017": [
-		-300.1296691894531,
-		172.68365478515625,
-		0.0
-	],
-	"g11018": [
-		-300.1579284667969,
-		-0.65093994140625,
-		0.0
-	],
-	"g11021": [
-		-300.3565979003906,
-		-520.3690185546875,
-		0.0
-	],
-	"g11113": [
-		-151.10986328125,
-		777.5980834960938,
-		0.0
-	],
-	"g11114": [
-		-151.134765625,
-		604.4125366210938,
-		0.0
-	],
-	"g11115": [
-		-150.3494873046875,
-		432.28802490234375,
-		0.0
-	],
-	"g11117": [
-		-150.12872314453125,
-		86.0755615234375,
-		0.0
-	],
-	"g11118": [
-		-150.0880126953125,
-		-87.24432373046875,
-		0.0
-	],
-	"g11119": [
-		-150.08807373046875,
-		-260.44927978515625,
-		0.0
-	],
-	"g11120": [
-		-150.0880126953125,
-		-433.65423583984375,
-		0.0
-	],
-	"g11121": [
-		-150.0880126953125,
-		-606.8593139648438,
-		0.0
-	],
-	"g11211": [
-		-1.10986328125,
-		1037.405517578125,
-		0.0
-	],
-	"g11212": [
-		-1.1085205078125,
-		864.1998901367188,
-		0.0
-	],
-	"g11213": [
-		-1.109619140625,
-		690.9955444335938,
-		0.0
-	],
-	"g11214": [
-		-1.109619140625,
-		517.7904663085938,
-		0.0
-	],
-	"g11215": [
-		-1.109619140625,
-		344.58538818359375,
-		0.0
-	],
-	"g11217": [
-		-0.0880126953125,
-		-0.64178466796875,
-		0.0
-	],
-	"g11218": [
-		-0.08807373046875,
-		-173.84686279296875,
-		0.0
-	],
-	"g11219": [
-		-0.088134765625,
-		-347.05194091796875,
-		0.0
-	],
-	"g11220": [
-		-0.088134765625,
-		-520.2568359375,
-		0.0
-	],
-	"g11221": [
-		-0.08819580078125,
-		-693.4619140625,
-		0.0
-	],
-	"g11311": [
-		148.890625,
-		950.8032836914062,
-		0.0
-	],
-	"g11312": [
-		148.8905029296875,
-		777.5982055664062,
-		0.0
-	],
-	"g11313": [
-		148.8905029296875,
-		604.3931274414062,
-		0.0
-	],
-	"g11314": [
-		149.07034301757812,
-		433.2586975097656,
-		1.862645149230957e-09
-	],
-	"g11315": [
-		149.07029724121094,
-		260.05364990234375,
-		1.862645149230957e-09
-	],
-	"g11316": [
-		149.0074005126953,
-		87.91712951660156,
-		0.0
-	],
-	"g11317": [
-		149.9119873046875,
-		-87.24432373046875,
-		0.0
-	],
-	"g11318": [
-		149.911865234375,
-		-260.44940185546875,
-		0.0
-	],
-	"g11319": [
-		149.91180419921875,
-		-433.65447998046875,
-		0.0
-	],
-	"g11320": [
-		149.91156005859375,
-		-606.8595581054688,
-		0.0
-	],
-	"g11413": [
-		299.0704040527344,
-		519.8612670898438,
-		0.0
-	],
-	"g11414": [
-		299.0703430175781,
-		346.65618896484375,
-		1.862645149230957e-09
-	],
-	"g11415": [
-		295.99090576171875,
-		171.0111846923828,
-		1.862645149230957e-09
-	],
-	"g11416": [
-		299.912109375,
-		-0.6416015625,
-		0.0
-	],
-	"g11417": [
-		299.9119873046875,
-		-173.84674072265625,
-		0.0
-	],
-	"g11418": [
-		299.911865234375,
-		-347.05194091796875,
-		0.0
-	],
-	"g11419": [
-		299.9117431640625,
-		-520.2570190429688,
-		0.0
-	],
-	"g11420": [
-		299.91552734375,
-		-693.4678344726562,
-		0.0
-	],
-	"g11421": [
-		299.8963623046875,
-		-866.6572875976562,
-		0.0
-	],
-	"g11512": [
-		449.0704345703125,
-		606.4638061523438,
-		-4.0046870708465576e-07
-	],
-	"g11513": [
-		449.07037353515625,
-		433.25872802734375,
-		1.862645149230957e-09
-	],
-	"g11514": [
-		449.0703125,
-		260.05364990234375,
-		1.862645149230957e-09
-	],
-	"g11515": [
-		451.3209228515625,
-		86.68668365478516,
-		4.6566128730773926e-09
-	],
-	"g11516": [
-		450.8204345703125,
-		-86.2853012084961,
-		0.0
-	],
-	"g11517": [
-		449.78082275390625,
-		-259.92742919921875,
-		0.0
-	],
-	"g11519": [
-		449.90869140625,
-		-606.8829345703125,
-		0.0
-	],
-	"g11520": [
-		449.8963623046875,
-		-780.0547485351562,
-		0.0
-	],
-	"g11521": [
-		449.8963623046875,
-		-953.2598266601562,
-		0.0
-	],
-	"g11612": [
-		599.0704345703125,
-		519.8540649414062,
-		-4.284083843231201e-08
-	],
-	"g11613": [
-		599.0685424804688,
-		346.6540222167969,
-		-4.284083843231201e-08
-	],
-	"g11614": [
-		605.47705078125,
-		180.02053833007812,
-		6.146728992462158e-08
-	],
-	"g11615": [
-		600.9014282226562,
-		-0.23164868354797363,
-		4.6938657760620117e-07
-	],
-	"g11616": [
-		600.16015625,
-		-172.87147521972656,
-		0.0
-	],
-	"g11617": [
-		600.22900390625,
-		-346.29901123046875,
-		0.0
-	],
-	"g11619": [
-		599.8966064453125,
-		-693.4522094726562,
-		0.0
-	],
-	"g11620": [
-		599.8968505859375,
-		-866.6596069335938,
-		0.0
-	],
-	"g11711": [
-		749.0704345703125,
-		606.4566040039062,
-		-4.284083843231201e-08
-	],
-	"g11712": [
-		755.470947265625,
-		438.18072509765625,
-		6.146728992462158e-08
-	],
-	"g11713": [
-		755.070068359375,
-		265.11236572265625,
-		6.146728992462158e-08
-	],
-	"g11714": [
-		755.233154296875,
-		93.22415161132812,
-		6.146728992462158e-08
-	],
-	"g11716": [
-		750.253173828125,
-		-259.66363525390625,
-		0.0
-	],
-	"g11717": [
-		750.29931640625,
-		-432.89617919921875,
-		0.0
-	],
-	"g11811": [
-		899.5013427734375,
-		529.2744140625,
-		7.636845111846924e-08
-	],
-	"g11812": [
-		899.5413208007812,
-		355.6947937011719,
-		6.146728992462158e-08
-	],
-	"g11813": [
-		905.233154296875,
-		179.82669067382812,
-		6.146728992462158e-08
-	],
-	"g11816": [
-		900.1815185546875,
-		-346.137451171875,
-		0.0
-	],
-	"g11911": [
-		1049.720703125,
-		442.1005859375,
-		1.6763806343078613e-08
-	],
-	"g12010": [
-		1201.000732421875,
-		525.80029296875,
-		-4.6566128730773926e-08
-	],
-	"g31118": [
-		-150.0880126953125,
-		-87.24432373046875,
-		0.0
-	],
-	"g31119": [
-		-150.08807373046875,
-		-260.44927978515625,
-		0.0
-	],
-	"g31120": [
-		-150.0880126953125,
-		-433.65423583984375,
-		0.0
-	],
-	"g31121": [
-		-150.0880126953125,
-		-606.8593139648438,
-		0.0
-	],
-	"g31217": [
-		-0.0880126953125,
-		-0.64178466796875,
-		0.0
-	],
-	"g31218": [
-		-0.08807373046875,
-		-173.84686279296875,
-		0.0
-	],
-	"g31219": [
-		-0.088134765625,
-		-347.05194091796875,
-		0.0
-	],
-	"g31220": [
-		-0.088134765625,
-		-520.2568359375,
-		0.0
-	],
-	"g31221": [
-		-0.08781299740076065,
-		-693.4619750976562,
-		0.0
-	],
-	"g31317": [
-		149.9119873046875,
-		-87.24432373046875,
-		0.0
-	],
-	"g31318": [
-		149.911865234375,
-		-260.44940185546875,
-		0.0
-	],
-	"g31319": [
-		149.91180419921875,
-		-433.65447998046875,
-		0.0
-	],
-	"g31320": [
-		149.91156005859375,
-		-606.8595581054688,
-		0.0
-	],
-	"g31416": [
-		299.912109375,
-		-0.6416015625,
-		0.0
-	],
-	"g31417": [
-		299.9119873046875,
-		-173.84674072265625,
-		0.0
-	],
-	"g31418": [
-		299.911865234375,
-		-347.05194091796875,
-		0.0
-	],
-	"g31419": [
-		299.9117431640625,
-		-520.2570190429688,
-		0.0
-	],
-	"g31518": [
-		449.9117431640625,
-		-433.6544494628906,
-		0.0
-	],
-	"g50100": [
-		0.4574565887451172,
-		-0.25896692276000977,
-		0.0
-	],
-	"g50200": [
-		-0.03338015079498291,
-		-0.046170711517333984,
-		0.0
-	],
-	"g50300": [
-		0.03852522373199463,
-		-0.06927397847175598,
-		0.016518428921699524
-	],
-	"g52001": [
-		-0.2926056385040283,
-		-0.7865628004074097,
-		0.0
-	],
-	"g53001": [
-		0.0,
-		0.0,
-		0.0
-	]
-}
