--- col/exporter/col_exporter.py
@@ -1,38 +1,0 @@
-from ...utils.ioUtils import write_uInt32
-from .col_colTreeNodes import write_col_colTreeNodes
-from .col_generate_data import COL_Data
-from .col_header import write_col_header
-from .col_meshes import write_col_meshes
-from .col_namegroups import write_col_namegroups
-
-
-def main(filepath, generateColTree):
-    data = COL_Data(generateColTree)
-
-    print('Creating col file: ', filepath)
-    col_file = open(filepath, 'wb')
-
-    print("Writing Header:")
-    write_col_header(col_file, data)
-
-    print("Writing NameGroups:")
-    write_col_namegroups(col_file, data)
-
-    print("Writing Meshes & Batches...")
-    write_col_meshes(col_file, data)
-
-    print("Writing Mesh & Bone Maps...")
-    if data.meshMapCount > 0:
-        col_file.seek(data.offsetMeshMap)
-        for num in data.meshMap:
-            write_uInt32(col_file, num)
-    data.boneMap.writeToFile(data.offsetBoneMap, col_file)
-    data.boneMap2.writeToFile(data.offsetBoneMap2, col_file)
-
-    print("Writing ColTreeNodes...")
-    write_col_colTreeNodes(col_file, data)
-
-    print("Finished exporting", filepath, "\nGood luck! :S")
-
-    col_file.flush()
-    col_file.close()
