+++ dat_dtt/exporter/dat_dtt_ui_manager.py
@@ -24,22 +24,6 @@ class ExportAllSteps(bpy.types.PropertyGroup):
         name = "Export WTA",
         default = True
     )
-    useColStep: bpy.props.BoolProperty(
-        name = "Export COL",
-        default = False
-    )
-    useLayStep: bpy.props.BoolProperty(
-        name = "Export LAY",
-        default = False
-    )
-    useSarStep: bpy.props.BoolProperty(
-        name = "Export SAR",
-        default = False
-    )
-    useGaStep: bpy.props.BoolProperty(
-        name = "Export GAArea",
-        default = False
-    )
     useDatStep: bpy.props.BoolProperty(
         name = "Export DAT",
         default = True
@@ -247,14 +231,6 @@ class ExportAll(bpy.types.Operator):
                 wmbFilePath = item.filepath
             elif item.filepath.endswith('.wta'):
                 wtaFilePath = item.filepath
-            elif item.filepath.endswith('.col'):
-                colFilePath = item.filepath
-            elif item.filepath.endswith('Layout.lay'):
-                layFilePath = item.filepath
-            elif item.filepath.endswith("GAArea.bxm"):
-                gaFilePath = item.filepath
-            elif item.filepath.endswith('.sar'):
-                sarFilePath = item.filepath
 
         for item in context.scene.DttContents:
             if item.filepath.endswith('.wtp'):
@@ -284,32 +260,6 @@ class ExportAll(bpy.types.Operator):
         if exportSteps.useWtpStep:
             print("Exporting WTP")
             export_wtp.main(context, wtpFilePath)
-            exportedFilesCount += 1
-        from ...col.exporter import col_exporter
-        if exportSteps.useColStep:
-            print("Exporting COL")
-            if exportSteps.triangulateMeshes:
-                triangulate_meshes("COL")
-            if exportSteps.centerOrigins:
-                centre_origins("COL")
-            if exportSteps.deleteLoose:
-                bpy.ops.b2n.deleteloosegeometryall()
-            col_exporter.main(colFilePath, True)
-            exportedFilesCount += 1
-        from ...lay.exporter import lay_exporter
-        if exportSteps.useLayStep:
-            print("Exporting LAY")
-            lay_exporter.main(layFilePath)
-            exportedFilesCount += 1
-        if exportSteps.useSarStep:
-            print("Exporting SAR")
-            from ...bxm.exporter import sarExporter
-            sarExporter.exportSar(sarFilePath)
-            exportedFilesCount += 1
-        if exportSteps.useGaStep:
-            print("Exporting GaArea")
-            from ...bxm.exporter import gaAreaExporter
-            gaAreaExporter.exportGaArea(gaFilePath)
             exportedFilesCount += 1
         from . import export_dat
         if exportSteps.useDatStep:
