--- sync/install_dependencies.py
@@ -1,192 +1,0 @@
-#    Copyright (C) 2020  Robert Guetzkow
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
-#    along with this program.  If not, see <https://www.gnu.org/licenses/>
-
-import traceback
-from typing import Tuple
-import bpy
-import os
-import sys
-import subprocess
-import importlib
-from collections import namedtuple
-
-from .shared import dependencies_installed, setDropDownOperatorAndIcon, removeDropDownOperatorAndIcon
-from ..utils.util import drawMultilineLabel
-from ..__init__ import bl_info
-
-Dependency = namedtuple("Dependency", ["module", "package", "name"])
-
-# Declare all modules that this add-on depends on, that may need to be installed. The package and (global) name can be
-# set to None, if they are equal to the module name. See import_module and ensure_and_import_module for the explanation
-# of the arguments. DO NOT use this to import other parts of your Python add-on, import them as usual with an
-# "import" statement.
-dependencies = (Dependency(module="websocket", package="websocket-client", name=None),)
-
-
-def import_module(module_name, global_name=None, reload=True):
-    """
-    Import a module.
-    :param module_name: Module to import.
-    :param global_name: (Optional) Name under which the module is imported. If None the module_name will be used.
-       This allows to import under a different name with the same effect as e.g. "import numpy as np" where "np" is
-       the global_name under which the module can be accessed.
-    :raises: ImportError and ModuleNotFoundError
-    """
-    if global_name is None:
-        global_name = module_name
-
-    if global_name in globals():
-        importlib.reload(globals()[global_name])
-    else:
-        # Attempt to import the module and assign it to globals dictionary. This allow to access the module under
-        # the given name, just like the regular import would.
-        globals()[global_name] = importlib.import_module(module_name)
-
-
-def install_pip():
-    """
-    Installs pip if not already present. Please note that ensurepip.bootstrap() also calls pip, which adds the
-    environment variable PIP_REQ_TRACKER. After ensurepip.bootstrap() finishes execution, the directory doesn't exist
-    anymore. However, when subprocess is used to call pip, in order to install a package, the environment variables
-    still contain PIP_REQ_TRACKER with the now nonexistent path. This is a problem since pip checks if PIP_REQ_TRACKER
-    is set and if it is, attempts to use it as temp directory. This would result in an error because the
-    directory can't be found. Therefore, PIP_REQ_TRACKER needs to be removed from environment variables.
-    :return:
-    """
-
-    try:
-        # Check if pip is already installed
-        subprocess.run([sys.executable, "-m", "pip", "--version"], check=True)
-    except subprocess.CalledProcessError:
-        import ensurepip
-
-        ensurepip.bootstrap()
-        os.environ.pop("PIP_REQ_TRACKER", None)
-
-
-def install_and_import_module(module_name, package_name=None, global_name=None):
-    """
-    Installs the package through pip and attempts to import the installed module.
-    :param module_name: Module to import.
-    :param package_name: (Optional) Name of the package that needs to be installed. If None it is assumed to be equal
-       to the module_name.
-    :param global_name: (Optional) Name under which the module is imported. If None the module_name will be used.
-       This allows to import under a different name with the same effect as e.g. "import numpy as np" where "np" is
-       the global_name under which the module can be accessed.
-    :raises: subprocess.CalledProcessError and ImportError
-    """
-    if package_name is None:
-        package_name = module_name
-
-    if global_name is None:
-        global_name = module_name
-
-    # Blender disables the loading of user site-packages by default. However, pip will still check them to determine
-    # if a dependency is already installed. This can cause problems if the packages is installed in the user
-    # site-packages and pip deems the requirement satisfied, but Blender cannot import the package from the user
-    # site-packages. Hence, the environment variable PYTHONNOUSERSITE is set to disallow pip from checking the user
-    # site-packages. If the package is not already installed for Blender's Python interpreter, it will then try to.
-    # The paths used by pip can be checked with `subprocess.run([bpy.app.binary_path_python, "-m", "site"], check=True)`
-
-    # Create a copy of the environment variables and modify them for the subprocess call
-    environ_copy = dict(os.environ)
-    environ_copy["PYTHONNOUSERSITE"] = "1"
-
-    subprocess.run([sys.executable, "-m", "pip", "install", package_name], check=True, env=environ_copy)
-
-    # The installation succeeded, attempt to import the module again
-    import_module(module_name, global_name)
-
-class EXAMPLE_OT_install_dependencies(bpy.types.Operator):
-    bl_idname = "sync.install_dependencies"
-    bl_label = "Sync First Time Setup"
-    bl_description = ("Needed for syncing with F-SERVO."
-                      "Downloads and installs the required python packages for this add-on. "
-                      "Internet connection is required. Blender may have to be started with "
-                      "elevated permissions in order to install the package")
-    bl_options = {"REGISTER", "INTERNAL"}
-
-    @classmethod
-    def poll(self, context):
-        # Deactivate when dependencies have been installed
-        return not dependencies_installed
-
-    def execute(self, context):
-        try:
-            install_pip()
-            for dependency in dependencies:
-                install_and_import_module(module_name=dependency.module,
-                                          package_name=dependency.package,
-                                          global_name=dependency.name)
-        except (subprocess.CalledProcessError, ImportError) as err:
-            traceback.print_exc()
-            self.report({"ERROR"}, "Failed to install dependencies! Try to restart Blender with elevated permissions")
-            return {"CANCELLED"}
-
-        global dependencies_installed
-        dependencies_installed = True
-
-        # Register the panels, operators, etc. since dependencies are installed
-        from .syncUi import registerSync
-        registerSync()
-        
-        removeDropDownOperatorAndIcon(getDropDownButton)
-
-        self.report({"INFO"}, "Setup complete. Ready to sync")
-
-        return {"FINISHED"}
-
-
-preference_classes = (
-    EXAMPLE_OT_install_dependencies,
-)
-
-def getDropDownButton() -> Tuple[str, str]:
-    return (
-        EXAMPLE_OT_install_dependencies.bl_idname,
-        "UV_SYNC_SELECT"
-    )
-
-def register():
-    global dependencies_installed
-    dependencies_installed = False
-
-    for cls in preference_classes:
-        bpy.utils.register_class(cls)
-
-    try:
-        for dependency in dependencies:
-            import_module(module_name=dependency.module, global_name=dependency.name)
-        dependencies_installed = True
-        from .syncUi import registerSync
-        registerSync()
-    except ModuleNotFoundError:
-        # Don't register other panels, operators etc.
-        pass
-    
-    if not dependencies_installed:
-        setDropDownOperatorAndIcon(getDropDownButton)
-
-
-def unregister():
-    for cls in preference_classes:
-        bpy.utils.unregister_class(cls)
-
-    if dependencies_installed:
-        from .syncUi import unregisterSync
-        unregisterSync()
-    
-    if not dependencies_installed:
-        removeDropDownOperatorAndIcon(getDropDownButton)
